Move the easy stuff out of LLM main
