+++ common/builders.py
@@ -1,7 +1,38 @@
 from composer import algorithms
+from composer.loggers import WandBLogger
+from composer.optim import DecoupledAdamW
+from composer.optim.scheduler import (ConstantWithWarmupScheduler,
+                                      CosineAnnealingWithWarmupScheduler)
+
+
+def build_logger(name, kwargs):
+    if name == 'wandb':
+        return WandBLogger(**kwargs)
+    else:
+        raise ValueError(f'Not sure how to build logger: {name}')
 
 def build_algorithm(name, kwargs):
     if name == 'gradient_clipping':
         return algorithms.GradientClipping(**kwargs)
     else:
         raise ValueError(f'Not sure how to build algorithm: {name}')
+
+def build_optimizer(cfg, model):
+    if cfg.name == 'decoupled_adamw':
+        return DecoupledAdamW(model.parameters(),
+                              lr=cfg.lr,
+                              betas=cfg.betas,
+                              eps=cfg.eps,
+                              weight_decay=cfg.weight_decay)
+    else:
+        raise ValueError(f'Not sure how to build optimizer: {cfg.name}')
+
+
+def build_scheduler(cfg):
+    if cfg.name == 'constant_with_warmup':
+        return ConstantWithWarmupScheduler(t_warmup=cfg.t_warmup)
+    elif cfg.name == 'cosine_with_warmup':
+        return CosineAnnealingWithWarmupScheduler(t_warmup=cfg.t_warmup,
+                                                  alpha_f=cfg.alpha_f)
+    else:
+        raise ValueError(f'Not sure how to build scheduler: {cfg.name}')
